pyproject.toml
[project]
name = "db-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx==0.27.2",
    "langchain==0.1.6",
    "langchain-community==0.0.20",
    "langchain-core==0.1.23",
    "langchain-experimental==0.0.49",
    "langchain-openai>=0.0.6",
    "openai==1.12.0",
    "pandas==2.2.2",
    "pyodbc==5.1.0",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.30",
    "streamlit>=1.41.1",
    "tabulate==0.9.0",
]
